v3: vindex: allow incompatible input keys
Issue #3151
The numeric vindexes were too strict about requiring number inputs.
This does not work correctly when results of joins are used as
vindex keys. Sometimes, they can be null, or may have invalid types.

In such cases, we should treat them as 'no match' instead of causing
an error. The only time a failure to match should result in an error
is on an insert. At other times, we should just return no results.