v3: extended vindexes.Lookup to require Update
With the Update function now part of Lookup, such vindexes can
now find more efficient implementations than the previous
Delete and Create sequence.

Currently, the Update implementations just do the Delete and Create,
but we'll soon improve those.

This abstraction also allows some vindexes to reject Updates if it's
not supported.

While making changes, I found a couple of bugs that I fixed:
* updateChangedVindexes fromIds computation was incorrect. It would
  have fetched incorrect values if the table had multiple multi-column
  and owned lookup vindexes.
* The Values for changing columns in planbuilder was getting filled
  out in the SET order, and not in the column order of the vindex,
  which was causing the updates to set the wrong values.

The above are not practical use cases, but it's better to strive for
correctness.