v3: support for Lookup vindex autocommit
Added autocommit support for lookup vindex. The flag is actually
autocommit_on_insert, because we autocommit only on insert. This
is because this the only operation where we stand to benefit
transactionally:

For Delete, we have to perform a select for update from the original
query, and then delete the vindexes, and then the actual row gets
deleted. So, it's hard to sequence autocommit transactions correctly
in such cases. But the natural commit order will usually be favorable.

For Update, it's even more non-trivial.